1.) bumped min julia version to 1.10 2.) Removed Requires.jl with the built-in support for package extensions introduced in version 1.9 3.) changed ci/cd yaml min version to 1.10
