+++ Project.toml
@@ -12,10 +12,25 @@ JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 MIMEs = "6c6e2e6c-3030-632d-7369-2d6c69616d65"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 RelocatableFolders = "05181044-ff0b-4ac5-8273-598c1e38db00"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
+
+[weakdeps]
+Bonito = "824d6782-a2ef-11e9-3a09-e5662e0c26f8"
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+Mustache = "ffc61752-8dc7-55ee-8c37-f3e9cdd09e70"
+OteraEngine = "b2d7f28f-acd6-4007-8b26-bc27716e5513"
+ProtoBuf = "3349acd9-ac6a-5e09-bcdb-63829b23a429"
+WGLMakie = "276b4fcb-3e11-5398-bf8b-a0c2d153d008"
+
+[extensions]
+BonitoExt = "Bonito"
+CairoMakieExt = "CairoMakie"
+MustacheExt = "Mustache"
+OteraEngineExt = "OteraEngine"
+ProtoBufExt = "ProtoBuf"
+WGLMakieExt = ["WGLMakie", "Bonito"]
 
 [compat]
 Bonito = "^3"
@@ -31,25 +46,18 @@ Pkg = "^1"
 ProtoBuf = "^1"
 Reexport = "^1"
 RelocatableFolders = "^1"
-Requires = "^1"
 Sockets = "^1"
 Statistics = "^1"
 StructTypes = "^1"
 Suppressor = "0.2.6, 1"
 WGLMakie = "^0.9"
-julia = "^1.6"
+julia = "^1.10"
 
 [extras]
-Bonito = "824d6782-a2ef-11e9-3a09-e5662e0c26f8"
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
-Mustache = "ffc61752-8dc7-55ee-8c37-f3e9cdd09e70"
-OteraEngine = "b2d7f28f-acd6-4007-8b26-bc27716e5513"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-ProtoBuf = "3349acd9-ac6a-5e09-bcdb-63829b23a429"
 StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-WGLMakie = "276b4fcb-3e11-5398-bf8b-a0c2d153d008"
 
 [targets]
-test = ["Bonito", "CairoMakie", "Mustache", "OteraEngine", "Pkg", "ProtoBuf", "StructTypes", "Suppressor", "Test", "WGLMakie"]
+test = ["Bonito", "CairoMakie", "Mustache", "OteraEngine", "ProtoBuf", "Pkg", "StructTypes", "Suppressor", "Test", "WGLMakie"]
