+++ ext/BonitoExt.jl
@@ -1,8 +1,12 @@
+module BonitoExt
+
 import HTTP
-import .Core.Util: html # import the html function from util so we can override it
-import .Bonito: Page, App
+import Oxygen.Util: html # import the html function from util so we can override it
+import Bonito: Page, App
 
 export html
+
+const HTML  = MIME"text/html"()
 
 """
 Converts a Figure object to the designated MIME type and wraps it inside an HTTP response.
@@ -23,10 +27,11 @@ function response(content::App, mime_type::MIME, status::Int, headers::Vector)
     return resp
 end
 
-
 """
     html(app::Bonito.App) :: HTTP.Response
 
 Convert a Bonito.App to HTML and wrap it inside an HTTP response.
 """
 html(app::App, status=200, headers=[]) :: HTTP.Response = response(app, HTML, status, headers)
+
+end
