+++ src/exts.jl
@@ -1,0 +1,19 @@
+
+export ProtoBuffer, protobuf
+export mustache, otera
+export png, svg, pdf
+
+# Serialization extension definitions
+function protobuf end
+struct ProtoBuffer{T} <: Extractor{T}
+    payload::T
+end
+
+# Templating extension definitions
+function mustache end
+function otera end
+
+# Plotting extension definitions
+function png end
+function svg end
+function pdf end
